Implement esapi crate and conf edition 2018
There are a number of changes in here:

* Implement the ESAPI wrapper crate
* Set up a simple connection context to the TPM
* Remove extern crate calls for edition 2018
* Comment out unused tpm calls in response handlers

Use of this change will require setting up the tpm2-tss. I will
document this in the readme and / or wiki
