+++ Cargo.toml
@@ -1,22 +1,24 @@
 [package]
+authors = ["Keylime Authors"]
+edition = "2018"
 name = "keylime_agent"
 version = "0.1.0"
-authors = ["Keylime Authors"]
 
 [dependencies]
-futures = "0.1.24"
-hyper = "0.12"
-rust-ini = "0.12.1"
-tokio-io = "0.1"
-serde_json = "1.0"
-pretty_env_logger = "0.2.0"
-log = "0.4"
-libc = "0.2.43"
-serde = "1.0.80"
-serde_derive = "1.0.80"
-rustc-serialize = "0.3.24"
-tempfile = "3.0.4"
-openssl = "0.10.15"
-hex = "0.3.2"
 base64 = "0.9.3"
 flate2 = "1.0.4"
+futures = "0.1.24"
+hex = "0.3.2"
+hyper = "0.12"
+libc = "0.2.43"
+log = "0.4"
+openssl = "0.10.15"
+pretty_env_logger = "0.2.0"
+rust-ini = "0.12.1"
+rustc-serialize = "0.3.24"
+serde = "1.0.80"
+serde_derive = "1.0.80"
+serde_json = "1.0"
+tempfile = "3.0.4"
+tokio-io = "0.1"
+tss-esapi = "4.0.9-alpha.1"
